app/requirements.txt
kafka-python
clickhouse-connect
Flask